testing/fullstack-ecommerce/backend/requirements.txt
# Core Django
Django==4.2.8
djangorestframework==3.14.0
django-filter==23.5
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Async Tasks
celery==5.3.4
redis==5.0.1

# Configuration
django-environ==0.11.2

# API Documentation
drf-spectacular==0.27.0

# Media/Files
Pillow==10.1.0

# Payment Processing
stripe==7.7.0

# Development Tools
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
Faker==20.1.0
black==23.12.1
mypy==1.7.1
ruff==0.1.8
django-stubs==4.2.7
djangorestframework-stubs==3.14.5

# Type Checking
types-redis==4.6.0.11
types-requests==2.31.0.10
